pyproject.toml
[tool.poetry]
name = "tg-task-tracker"
version = "0.1.0"
description = ""
authors = ["MaksimEvdokimov <maksorli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg = {extras = ["binary"], version = "^3.2.1"}
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
